create internal test branch, restructure discovery code
