+++ .gitignore
@@ -131,3 +131,7 @@ dmypy.json
 
 # VuePress
 node_modules/
+
+local stuff
+.vscode/
+build.sh
