+++ TheengsGateway/ble_gateway.py
@@ -22,7 +22,6 @@
 # python 3.6
 
 import asyncio
-from distutils.command.config import config
 import json
 import struct
 import sys
@@ -32,14 +31,9 @@ from bleak import BleakScanner
 from ._decoder import decodeBLE, getProperties, getAttribute
 from paho.mqtt import client as mqtt_client
 from threading import Thread
+from .discovery import discovery as dc
 
 logger = logging.getLogger('BLEGateway')
-class homeassistant:
-    def __init__(self) -> None:
-        self.device = {}
-        self.device['name'] = "BLEGateway"
-        self.device['identifiers'] = "BLEGateway"
-        self.device['manufacturer'] = "theengs"
 
 class gateway:
     def __init__(self, broker, port, username, password, discovery):
@@ -87,46 +81,7 @@ class gateway:
         else:
             logger.error(f"Failed to send message to topic {pub_topic}")
     
-    def publish_device_info(self, pub_device):  ##publish sensor directly to home assistant via mqtt discovery
-        ha = homeassistant()
-        device = {}
-        ##setup HA device
-        pub_device = json.loads(pub_device)
-        pub_device_uuid = pub_device['id']
-        pub_device_uuid = pub_device_uuid.replace(':', '')
-        device['unique_id'] = pub_device['id']
-        topic = self.discovery_topic + "/" + pub_device_uuid
-        state_topic = topic + "/state"
-        config_topic = topic + "/config"
-        attr_topic = topic + "/attributes"
-        if 'name' in pub_device:
-          device['name'] = pub_device['name']
-        else: 
-            device['name'] = pub_device_uuid
-        device['state_topic'] = state_topic
-        device['device'] = ha.device
-        device['schema'] = "json"
-        device['state_topic'] = state_topic
-        if 'steps' in pub_device:  ## sensor states for Xaomi Mi Band
-          msg = pub_device['steps']
-          self.publish(msg, state_topic) 
-        
-        ##attributes
-        attributes = {}      
-        attributes['rssi'] = pub_device['rssi']
-        attributes['brand'] = pub_device['brand']
-        attributes['id'] = pub_device['id']
-        attributes['model'] = pub_device['model']
-        attributes['model_id'] = pub_device['model_id']
-        attributes = json.dumps(attributes)
-        device['json_attr'] = attributes
-        device['json_attr_t'] = attr_topic
-        self.publish(attributes, attr_topic) ##attributes
-        
-        
-        payload = json.dumps(device)
-        msg = payload
-        self.publish(msg, config_topic) ##overall device
+
 
     async def ble_scan_loop(self):
         scanner = BleakScanner()
@@ -189,7 +144,7 @@ def run(arg):
         raise SystemExit(f"Invalid File: {sys.argv[1]}")
 
     try:
-        gw = gateway(config["host"], int(config["port"]), config["user"], config["pass"], config["discovery"])
+        gw = dc(config["host"], int(config["port"]), config["user"], config["pass"], config["discovery"])
     except:
         raise SystemExit(f"Missing or invalid MQTT host parameters")
